v1.6.12 - Fixed problem with using fault in terrathread (empty codeseg) - Fixed fault to properly count faultNum and added unsafe mode - Moved update delay counter into new class called Synchroniser
