+++ TeamCode/src/main/java/util/TerraThread.java
@@ -12,18 +12,15 @@ public class TerraThread extends Thread {
 
     private volatile Status currentStatus = Status.ACTIVE;
     private volatile CodeSeg updateCode = () -> {};
-    private volatile boolean ready = false;
 
     public synchronized void setCode(CodeSeg cs){
         updateCode = cs;
-        ready = true;
     }
 
     public synchronized void stopUpdating(){currentStatus = Status.DISABLED;}
 
     @Override
     public void run() {
-        fault.check("Ran thread with null code", Expectation.UNEXPECTED, Magnitude.MODERATE, ready);
         while (!currentStatus.equals(Status.DISABLED)){
             if(currentStatus.equals(Status.IDLE)){
                 synchronized (this) {
